`<Page/>` - move e2e examples to separate stories (#1940)
