+++ .storybook/stories.js
@@ -124,3 +124,6 @@ import '../stories/TPA/Label';
 // DEPRECATED
 import '../stories/LanguagePicker/index.story.js';
 import '../stories/SideBar';
+
+// Tests 
+import '../stories/Page/PageTestStories.js'; // Tests/2.5 + Page/
