+++ stories/Page/storybookConfig.js
@@ -1,0 +1,4 @@
+export const storybookConfig = {
+  category: '2. Layout',
+  storyName: '2.5 Page'
+};
