+++ stories/storyCategories.js
@@ -1,0 +1,4 @@
+// We can add here constants for the Storybook category names
+
+// Root Level
+export const TESTS_PREFIX = 'Tests';
